#49 Added task option inheritance.
