+++ src/main/groovy/com/redpillanalytics/odi/gradle/OdiPlugin.groovy
@@ -7,6 +7,7 @@ import com.redpillanalytics.odi.gradle.tasks.CreateProjectTask
 import com.redpillanalytics.odi.gradle.tasks.DeleteProjectTask
 import com.redpillanalytics.odi.gradle.tasks.ExportLoadPlanDirectoryTask
 import com.redpillanalytics.odi.gradle.tasks.ExportModelDirectoryTask
+import com.redpillanalytics.odi.gradle.tasks.ExportObjectDirectoryTask
 import com.redpillanalytics.odi.gradle.tasks.ExportProjectDirectoryTask
 import com.redpillanalytics.odi.gradle.tasks.ExportProjectFileTask
 import com.redpillanalytics.odi.gradle.tasks.ImportDirectoryTask
@@ -186,21 +187,31 @@ class OdiPlugin implements Plugin<Project> {
                   outputs.upToDateWhen { false }
                }
 
+               project.task(bg.getTaskName('exportProjectDir'), type: ExportProjectDirectoryTask) {
+
+                  group taskGroup
+                  description """Run all project directory export tasks from ODI project '${defaultProjectCode}'${
+                     projectFolder ? " for folder '$projectFolder'" : ''
+                  }."""
+                  projectCode defaultProjectCode
+                  folderName projectFolder
+               }
+
                // enable tasks for all the different object types in a project
-               project.task(bg.getTaskName('exportMappingDir'), type: ExportProjectDirectoryTask) {
+               project.task(bg.getTaskName('exportMappingDir'), type: ExportObjectDirectoryTask) {
 
                   group taskGroup
                   description """Export mappings from ODI project '${defaultProjectCode}' ${
                      projectFolder ? "for folder '$projectFolder'" : ''
                   }from the ODI repository into source control."""
                   objectType 'mapping'
-                  projectCode defaultProjectCode
+                  projectCode project.tasks."${bg.getTaskName('exportMappingDir')}".projectCode ?: defaultProjectCode
+                  folderName project.tasks."${bg.getTaskName('exportMappingDir')}".folderName ?: projectFolder
                   instance odiInstance
-                  folderName projectFolder
                   outputs.upToDateWhen { false }
                }
 
-               project.task(bg.getTaskName('exportReusableMappingDir'), type: ExportProjectDirectoryTask) {
+               project.task(bg.getTaskName('exportReusableMappingDir'), type: ExportObjectDirectoryTask) {
 
                   group taskGroup
                   description """Export reusable mappings from ODI project '${defaultProjectCode}' ${
@@ -213,7 +224,7 @@ class OdiPlugin implements Plugin<Project> {
                   outputs.upToDateWhen { false }
                }
 
-               project.task(bg.getTaskName('exportProcedureDir'), type: ExportProjectDirectoryTask) {
+               project.task(bg.getTaskName('exportProcedureDir'), type: ExportObjectDirectoryTask) {
 
                   group taskGroup
                   description """Export procedures from ODI project '${defaultProjectCode}' ${
@@ -226,7 +237,7 @@ class OdiPlugin implements Plugin<Project> {
                   outputs.upToDateWhen { false }
                }
 
-               project.task(bg.getTaskName('exportPackageDir'), type: ExportProjectDirectoryTask) {
+               project.task(bg.getTaskName('exportPackageDir'), type: ExportObjectDirectoryTask) {
 
                   group taskGroup
                   description """Export packages from ODI project '${defaultProjectCode}' ${
@@ -239,14 +250,12 @@ class OdiPlugin implements Plugin<Project> {
                   outputs.upToDateWhen { false }
                }
 
-               project.task(bg.getTaskName('exportProjectDir')) {
+               project."${bg.getTaskName('exportProjectDir')}".dependsOn project."${bg.getTaskName('exportMappingDir')}",
+                       project."${bg.getTaskName('exportReusableMappingDir')}",
+                       project."${bg.getTaskName('exportProcedureDir')}",
+                       project."${bg.getTaskName('exportPackageDir')}"
 
-                  group taskGroup
-                  description """Run all project directory export tasks from ODI project '${defaultProjectCode}'${
-                     projectFolder ? " for folder '$projectFolder'" : ''
-                  }."""
-                  dependsOn bg.getTaskName('exportMappingDir'), bg.getTaskName('exportReusableMappingDir'), bg.getTaskName('exportProcedureDir'), bg.getTaskName('exportPackageDir')
-               }
+               //dependsOn bg.getTaskName('exportMappingDir'), bg.getTaskName('exportReusableMappingDir'), bg.getTaskName('exportProcedureDir'), bg.getTaskName('exportPackageDir')
 
                project."${bg.getTaskName('exportMappingDir')}".mustRunAfter project."${bg.getTaskName('exportReusableMappingDir')}"
                project."${bg.getTaskName('exportPackageDir')}".mustRunAfter project."${bg.getTaskName('exportProcedureDir')}"
